Temperature and graph axis demo
An alternative method for drawing the graphs on the figure using twinx(). In theory, we should be able to use this format, maybe make it a bit more organized, to add and manage more axes to the graph. Additionally, it seemed to require drawing the axis of the wind vectors last otherwise the y-limits would overlap with the AOD y-limits.