Added vector sum and averaging for Wind Vectors.
Used vector decomposition to break down the magnitude and angle of each wind vector into X and Y coordinates, appended those coordinates to the wind data frame (WNDdf) as columns 5 and 6 respectively. In order to change the sampling rate for the wind vectors either change the variable "windSampleRate" on line 13 to the desired rate or change the individual parts in the quiver calculations. Added a timestamp to the auto save feature (Lines 111-117).